+++ bcbio/ngsalign/postalign.py
@@ -19,6 +19,7 @@ from bcbio.log import logger
 from bcbio.pipeline import config_utils
 from bcbio.pipeline import datadict as dd
 from bcbio.provenance import do
+from bcbio.variation import coverage
 
 pysam = utils.LazyImport("pysam")
 
@@ -160,6 +161,25 @@ def _get_fgbio_jvm_opts(data, tmpdir, scale_factor=None):
     jvm_opts = " ".join(jvm_opts)
     return jvm_opts + " --tmp-dir %s" % tmpdir
 
+def _estimate_fgbio_defaults(avg_coverage):
+    """Provide fgbio defaults based on input sequence depth and coverage.
+
+    For higher depth/duplication we want to use `--min-reads` to allow
+    consensus calling in the duplicates:
+
+    https://fulcrumgenomics.github.io/fgbio/tools/latest/CallMolecularConsensusReads.html
+
+    If duplicated adjusted depth leaves a coverage of 800x or higher
+    (giving us ~4 reads at 0.5% detection frequency),
+    then we use `--min-reads 2`, otherwise `--min-reads 1`
+    """
+    out = {}
+    if avg_coverage >= 800:
+        out["--min-reads"] = 2
+    else:
+        out["--min-reads"] = 1
+    return out
+
 def umi_consensus(data):
     """Convert UMI grouped reads into fastq pair for re-alignment.
     """
@@ -167,12 +187,14 @@ def umi_consensus(data):
     umi_method, umi_tag = _check_umi_type(align_bam)
     f1_out = "%s-cumi-1.fq.gz" % utils.splitext_plus(align_bam)[0]
     f2_out = "%s-cumi-2.fq.gz" % utils.splitext_plus(align_bam)[0]
+    avg_coverage = coverage.get_average_coverage("rawumi", dd.get_variant_regions(data), data)
     if not utils.file_uptodate(f1_out, align_bam):
         with file_transaction(data, f1_out, f2_out) as (tx_f1_out, tx_f2_out):
             jvm_opts = _get_fgbio_jvm_opts(data, os.path.dirname(tx_f1_out), 2)
             # Improve speeds by avoiding compression read/write bottlenecks
             io_opts = "--async-io=true --compression=0"
-            group_opts, cons_opts, filter_opts = _get_fgbio_options(data, umi_method)
+            est_options = _estimate_fgbio_defaults(avg_coverage)
+            group_opts, cons_opts, filter_opts = _get_fgbio_options(data, est_options, umi_method)
             cons_method = "CallDuplexConsensusReads" if umi_method == "paired" else "CallMolecularConsensusReads"
             tempfile = "%s-bamtofastq-tmp" % utils.splitext_plus(f1_out)[0]
             ref_file = dd.get_ref_file(data)
@@ -185,7 +207,7 @@ def umi_consensus(data):
                    "-i /dev/stdin -o /dev/stdout | "
                    "bamtofastq collate=1 T={tempfile} F={tx_f1_out} F2={tx_f2_out} tags=cD,cM,cE gz=1")
             do.run(cmd.format(**locals()), "UMI consensus fastq generation")
-    return f1_out, f2_out
+    return f1_out, f2_out, avg_coverage, data
 
 def _check_umi_type(bam_file):
     """Determine the type of UMI from BAM tags: standard or paired.
@@ -205,7 +227,7 @@ def _check_umi_type(bam_file):
                 else:
                     return "adjacency", tag
 
-def _get_fgbio_options(data, umi_method):
+def _get_fgbio_options(data, estimated_defaults, umi_method):
     """Get adjustable, through resources, or default options for fgbio.
     """
     group_opts = ["--edits", "--min-map-q"]
@@ -220,6 +242,7 @@ def _get_fgbio_options(data, umi_method):
                 "--max-base-error-rate": "0.1",
                 "--min-input-base-quality": "2",
                 "--edits": "1"}
+    defaults.update(estimated_defaults)
     ropts = config_utils.get_resources("fgbio", data["config"]).get("options", [])
     assert len(ropts) % 2 == 0, "Expect even number of options for fgbio" % ropts
     ropts = dict(tz.partition(2, ropts))